requirements.txt
aiogram==3.7.0
apscheduler==3.10.4
httpx==0.27.0
feedparser==6.0.11
Pillow==10.3.0
pytz==2024.1
google-generativeai==0.7.2
